refactor register to multiple functions
